Hyscale commands to its sub-command package refactoring
