+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/commands/deploy/HyscaleDeployCommand.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.hyscale.controller.commands;
+package io.hyscale.controller.commands.deploy;
 
 import java.util.concurrent.Callable;
 
+import io.hyscale.controller.commands.HyscaleCommand;
 import org.springframework.stereotype.Component;
 
 import io.hyscale.commons.constants.ToolConstants;
@@ -24,7 +25,7 @@ import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
 /**
- * This class executes 'hyscale generate' command
+ * This class executes 'hyscale deploy' command
  * It is a sub-command of the 'hyscale' command
  * @see HyscaleCommand
  * Every command/sub-command has to implement the {@link Callable} so that
@@ -32,21 +33,22 @@ import picocli.CommandLine.Command;
  * method will be invoked
  *
  * The sub-commands of are handled by @Command annotation
+ *
  */
-@Command(name = "generate", subcommands = { HyscaleGenerateServiceCommand.class }, description = { "Generates the specified resource" })
+@Command(name = "deploy", subcommands = { HyscaleDeployServiceCommand.class}, description = "Deploys the specified resource")
 @Component
-public class HyscaleGenerateCommand implements Callable<Integer> {
+public class HyscaleDeployCommand implements Callable<Integer> {
 
 	@CommandLine.Option(names = {"-h", "--help"}, usageHelp = true, description = "Displays the help information of the specified command")
 	private boolean helpRequested = false;
 
 	/**
-	 * Executes the 'hyscale generate' command
+	 * Executes the 'hyscale deploy' command
 	 * Provides usage of this command to the user
 	 */
 	@Override
     public Integer call() throws Exception {
-		new CommandLine(new HyscaleGenerateCommand()).usage(System.out);
+		new CommandLine(new HyscaleDeployCommand()).usage(System.out);
 		return ToolConstants.INVALID_INPUT_ERROR_CODE;
 	}
 }
