+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/commands/generate/HyscaleGenerateCommand.java
@@ -13,34 +13,41 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.hyscale.controller.commands;
+package io.hyscale.controller.commands.generate;
 
 import java.util.concurrent.Callable;
 
+import io.hyscale.controller.commands.HyscaleCommand;
 import org.springframework.stereotype.Component;
 
 import io.hyscale.commons.constants.ToolConstants;
 import picocli.CommandLine;
+import picocli.CommandLine.Command;
 
 /**
- * This class executes 'hyscale generate service' command
- * It is a sub-command of the 'hyscale generate' command
- * @see HyscaleGenerateCommand
+ * This class executes 'hyscale generate' command
+ * It is a sub-command of the 'hyscale' command
+ * @see HyscaleCommand
  * Every command/sub-command has to implement the {@link Callable} so that
  * whenever the command is executed the {@link #call()}
  * method will be invoked
  *
  * The sub-commands of are handled by @Command annotation
- *
  */
-@CommandLine.Command(name = "service", subcommands = { HyscaleGenerateServiceManifestsCommand.class} ,
-        description = "Performs action on the service")
+@Command(name = "generate", subcommands = { HyscaleGenerateServiceCommand.class }, description = { "Generates the specified resource" })
 @Component
-public class HyscaleGenerateServiceCommand implements Callable<Integer> {
-    
-    @Override
+public class HyscaleGenerateCommand implements Callable<Integer> {
+
+	@CommandLine.Option(names = {"-h", "--help"}, usageHelp = true, description = "Displays the help information of the specified command")
+	private boolean helpRequested = false;
+
+	/**
+	 * Executes the 'hyscale generate' command
+	 * Provides usage of this command to the user
+	 */
+	@Override
     public Integer call() throws Exception {
-        new CommandLine(new HyscaleGenerateServiceCommand()).usage(System.out);
-        return ToolConstants.INVALID_INPUT_ERROR_CODE;
-    }
+		new CommandLine(new HyscaleGenerateCommand()).usage(System.out);
+		return ToolConstants.INVALID_INPUT_ERROR_CODE;
+	}
 }
