+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/commands/generate/HyscaleGenerateServiceManifestsCommand.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.hyscale.controller.commands;
+package io.hyscale.controller.commands.generate;
 
 import java.io.File;
 import java.util.List;
