+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/commands/get/app/HyscaleAppStatusCommand.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.hyscale.controller.commands;
+package io.hyscale.controller.commands.get.app;
 
 import java.util.ArrayList;
 import java.util.List;
