+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/commands/undeploy/HyscaleUndeployCommand.java
@@ -13,19 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.hyscale.controller.commands;
+package io.hyscale.controller.commands.undeploy;
 
 import java.util.concurrent.Callable;
 
+import io.hyscale.controller.commands.HyscaleCommand;
 import org.springframework.stereotype.Component;
 
 import io.hyscale.commons.constants.ToolConstants;
 import picocli.CommandLine;
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
 
 /**
- *  This class executes  'hyscale get service' command
- *  It is a sub-command of the 'hyscale get ' command
- *  @see HyscaleGetCommand .
+ *  This class executes 'hyscale undeploy' command
+ *  It is a sub-command of the 'hyscale' command
+ *  @see HyscaleCommand
  *  Every command/sub-command has to implement the {@link Callable} so that
  *  whenever the command is executed the {@link #call()}
  *  method will be invoked
@@ -33,21 +36,22 @@ import picocli.CommandLine;
  * The sub-commands of are handled by @Command annotation
  *
  */
-@CommandLine.Command(name = "service", subcommands = { HyscaleServiceLogsCommand.class,
-		HyscaleServiceStatusCommand.class }, description = "Performs action on the service")
+@Command(name = "undeploy", subcommands = {HyscaleUndeployAppCommand.class,
+		HyscaleUndeploySeviceCommand.class }, description = "Undeploys the specified resource")
 @Component
-public class HyscaleGetServiceCommand implements Callable<Integer> {
+public class HyscaleUndeployCommand implements Callable<Integer> {
 
-	@CommandLine.Option(names = {"-h", "--help"}, usageHelp = true, description = "Displays the help information of the specified command")
+	@Option(names = { "-h", "--help" }, usageHelp = true, description = "Display help message")
 	private boolean helpRequested = false;
 
 	/**
-	 * Executes the 'hyscale get service' command
+	 * Executes the 'hyscale undeploy' command
 	 * Provides usage of this command to the user
 	 */
 	@Override
     public Integer call() throws Exception {
-	    new CommandLine(new HyscaleGetServiceCommand()).usage(System.out);
+	    new CommandLine(new HyscaleUndeployCommand()).usage(System.out);
 	    return ToolConstants.INVALID_INPUT_ERROR_CODE;
 	}
+
 }
