+++ workflow-controller/controller-service/src/main/java/io/hyscale/controller/invoker/DeployComponentInvoker.java
@@ -200,7 +200,7 @@ public class DeployComponentInvoker extends ComponentInvoker<WorkflowContext> {
         }
         return null;
     }
-        
+
     /**
      * Write deployment logs to file for later access
      *
@@ -208,9 +208,11 @@ public class DeployComponentInvoker extends ComponentInvoker<WorkflowContext> {
      * @param deploymentContext
      */
     private void writeDeployLogs(WorkflowContext context, DeploymentContext deploymentContext) {
-        try (InputStream is = deployer.logs(deploymentContext)) {
+        String serviceName = deploymentContext.getServiceName();
+        try (InputStream is = deployer.logs(deploymentContext.getAuthConfig(), serviceName,
+                deploymentContext.getNamespace(), null, serviceName, deploymentContext.getReadLines(), deploymentContext.isTailLogs())) {
             String deploylogFile = deployerConfig.getDeployLogDir(deploymentContext.getAppName(),
-                    deploymentContext.getServiceName());
+                    serviceName);
             logProcessor.writeLogFile(is, deploylogFile);
             context.addAttribute(WorkflowConstants.DEPLOY_LOGS,
                     deployerConfig.getDeployLogDir(deploymentContext.getAppName(), deploymentContext.getServiceName()));
@@ -228,7 +230,7 @@ public class DeployComponentInvoker extends ComponentInvoker<WorkflowContext> {
         StringBuilder errorMessage = new StringBuilder();
         errorMessage.append(he != null ? he.getMessage() : DeployerErrorCodes.FAILED_TO_APPLY_MANIFEST.getErrorMessage());
         if (troubleshootMsgObj != null) {
-            String troubleshootMessage = (String)troubleshootMsgObj;
+            String troubleshootMessage = (String) troubleshootMsgObj;
             logger.error("Troubleshoot message: {}", troubleshootMessage);
             errorMessage.append(ToolConstants.NEW_LINE).append(troubleshootMessage);
         }
